# app/static/swagger.yml
swagger: "2.0"
info:
  title: Mechanic Shop API
  description: This is an API to manage a mechanic shop
  version: "1.0.0"
host: "mechanic-api-zfu6.onrender.com"
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: Login endpoint (generates token)
      description: Validates customer credentials, generates and return a customer specific token.
      parameters:
        - in: body
          name: body
          description: Login Credentials
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: Login successful
          schema:
            $ref: "#/definitions/LoginResponse"
  /customers:
    post:
      tags:
        - Customers
      summary: Endpoint to create/register a new customer
      description: Endpoint to create a new customer in our mechanic shop system, emails and phone numbers must be unique.
      parameters:
        - in: body
          name: body
          description: Information required to create a new customer
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: Customer created
          schema:
            $ref: "#/definitions/CreateCustomerResponse"

    get:
      tags:
        - Customers
      summary: Returns all customers
      description: Endpoint to retrieve a list of all customers
      responses:
        200:
          description: Retrieved customers successfully
          schema:
            $ref: "#/definitions/AllCustomers"

    put:
      tags:
        - Customers
      summary: Endpoint to update a customer
      description: Endpoint to update a customer, this is a token authenticated route, and the customer must be logged in to update their account
      consumes:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: Information requried to update a customer
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          descriptions: Successfully updated customer
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
        400:
          description: Validation error
        401:
          description: Unauthorized
        404:
          description: Customer not found

    delete:
      tags:
        - Customers
      summary: Endpoint to delete a customer
      description: Endpoint to delete a customer. Requires token authentication, customer must be logged in to delete their account
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successfully deleted customer
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: Endpoint to get a single customer
      description: Endpoint to retrieve a single customer
      parameters:
        - in: path
          name: customer_id
          description: ID of customer to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Retrieved customer successfully
          schema:
            $ref: "#/definitions/SingleCustomer"
...
definitions:
  LoginCredentials:
    type: object
    properties:
      customer_email:
        type: string
      customer_password:
        type: string

    required:
      - customer_email
      - customer_password
  LoginResponse:
    type: object
    properties:
      token:
        type: string
      message:
        type: string
      status:
        type: string
  CreateCustomerPayload:
    type: object
    properties:
      customer_name:
        type: string
      customer_email:
        type: string
      customer_password:
        type: string
      customer_phone:
        type: string

    required:
      - customer_name
      - customer_email
      - customer_password
      - customer_phone
  CreateCustomerResponse:
    type: object
    properties:
      id:
        type: integer
      customer_name:
        type: string
      customer_email:
        type: string
      customer_password:
        type: string
      customer_phone:
        type: string

    example:
      id: 1
      customer_name: "John Doe"
      customer_email: "jdoe@example.com"
      customer_password: "test1234"
      customer_phone: "1234567890"
  AllCustomers:
    type: array
    items:
      type: object
      properties:
        id:
          type: integer
        customer_name:
          type: string
        customer_email:
          type: string
        customer_password:
          type: string
        customer_phone:
          type: string
  UpdateCustomerPayload:
    type: object
    properties:
      customer_name:
        type: string
      customer_email:
        type: string
      customer_password:
        type: string
      customer_phone:
        type: string

    required:
      - customer_name
      - customer_email
      - customer_password
      - customer_phone
  UpdateCustomerResponse:
    type: object
    properties:
      id:
        type: integer
      customer_name:
        type: string
      customer_email:
        type: string
      customer_password:
        type: string
      customer_phone:
        type: string

    example:
      id: 1
      customer_name: "John Doe"
      customer_email: "johnd@example.com"
      customer_password: "test1234"
      customer_phone: "1234567890"
  DeleteCustomerResponse:
    type: object
    properties:
      message:
        type: string
    example:
      message: "successfully deleted customer <customer_id>"
  SingleCustomer:
    type: object
    properties:
      id:
        type: integer
      customer_name:
        type: string
      customer_email:
        type: string
      customer_password:
        type: string
      customer_phone:
        type: string
...